+++ src/js/components/options/OptionsDialog.js
@@ -1,41 +1,35 @@
 import React from 'react'
-import Dialog from 'material-ui/Dialog/Dialog'
+import { connect } from 'react-redux'
+
 import Transition from '../Transition'
+import LegacyToggle from './LegacyToggle'
+import TagMenu from './TagMenu'
+
+import Dialog from 'material-ui/Dialog/Dialog'
 import AppBar from 'material-ui/AppBar/AppBar'
 import Toolbar from 'material-ui/Toolbar/Toolbar'
 import Typography from 'material-ui/Typography'
 import IconButton from 'material-ui/IconButton'
 import Icon from 'material-ui/Icon'
-import TextField from 'material-ui/TextField'
-import Switch from 'material-ui/Switch'
 import Divider from 'material-ui/Divider/Divider'
-import { FormGroup, FormControlLabel, FormLabel } from 'material-ui/Form'
 
-import { connect } from 'react-redux'
-import { legacyToggle, prefixChange } from '../../actions/options'
-
+import { prefixChange } from '../../actions/options'
 import propTypes from 'prop-types'
 
 import '../../../styles/options.css'
 
 const mapStateToProps = (state) => ({
   options: {
-    localVarPrefix: state.options.localVarPrefix,
-    legacyMode: state.options.legacyMode
+    localVarPrefix: state.options.localVarPrefix
   }
 })
 
 const mapDispatchToProps = (dispatch) => ({
-  onLegacyChange: () => dispatch(legacyToggle()),
   onPrefixChange: (event) => dispatch(prefixChange(event.target.id, event.target.value))
 })
 
 const OptionsDialog = (props) => {
-  let { isOpen, options, onClose, onLegacyChange, onPrefixChange } = props
-
-  const handleTextFieldClick = (event) => {
-    event.stopPropagation()
-  }
+  let { isOpen, onClose } = props
 
   return (
     <Dialog 
@@ -58,28 +52,9 @@ const OptionsDialog = (props) => {
         </Toolbar>
       </AppBar>
       <div className="options-root">
-        <FormGroup>
-          <FormLabel component="legend">Documentation Style</FormLabel>
-          <FormControlLabel
-            control={
-              <Switch
-                checked={options.legacyMode}
-                onChange={onLegacyChange}
-              />
-            }
-            label={options.legacyMode ? 'GameMaker: Studio 1.4' : 'GameMaker Studio 2'}
-          />
-        </FormGroup>
+        <LegacyToggle/>
         <Divider/>
-        <TextField
-          id="localVarPrefix"
-          label="Local Variable Prefix"
-          placeholder="none"
-          margin="normal"
-          value={options.localVarPrefix}
-          onClick={handleTextFieldClick}
-          onChange={onPrefixChange}
-        />
+        <TagMenu/>
       </div>
     </Dialog>
   )